change init (check doublons and split the params)

+++ src/atoi.c
@@ -6,16 +6,16 @@
 /*   By: antoinemura <antoinemura@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/13 19:45:09 by marvin            #+#    #+#             */
-/*   Updated: 2024/12/11 01:53:15 by antoinemura      ###   ########.fr       */
+/*   Updated: 2024/12/11 13:07:19 by antoinemura      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-int	ft_atoi(const char *nptr)
+long long	ft_atoi(const char *nptr)
 {
-	int	signe;
-	int	total;
+	int			signe;
+	long long	total;
 
 	signe = 1;
 	total = 0;
